refactor: 양의 정수 조건 추가, 남아있는 int부분 long으로 변경

+++ src/main/java/controller/Validator.java
@@ -30,7 +30,7 @@ public class Validator {
 
     public static boolean isNumber(String startValue) {
         try {
-            Integer.parseInt(startValue);
+            Long.parseLong(startValue);
             return true;
         } catch (NumberFormatException e) {
             return false;
